Renamed folders in test-modules to match the applications structure
