+++ data-catalog-backend-test/data-catalog-backend-test-component/src/test/java/no/nav/data/catalog/backend/app/dataset/DatasetControllerTest.java
@@ -1,6 +1,10 @@
 package no.nav.data.catalog.backend.app.dataset;
 
 import no.nav.data.catalog.backend.app.AppStarter;
+import no.nav.data.catalog.backend.app.dataset.Dataset;
+import no.nav.data.catalog.backend.app.dataset.DatasetController;
+import no.nav.data.catalog.backend.app.dataset.DatasetData;
+import no.nav.data.catalog.backend.app.dataset.DatasetService;
 import no.nav.data.catalog.backend.app.dataset.repo.DatasetRepository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
