+++ data-catalog-backend-test/data-catalog-backend-test-component/src/test/java/no/nav/data/catalog/backend/app/github/GithubWebhooksControllerTest.java
@@ -6,6 +6,10 @@ import no.nav.data.catalog.backend.app.dataset.Dataset;
 import no.nav.data.catalog.backend.app.dataset.DatasetRequest;
 import no.nav.data.catalog.backend.app.dataset.DatasetService;
 import no.nav.data.catalog.backend.app.dataset.repo.DatasetRepository;
+import no.nav.data.catalog.backend.app.github.GithubConfig;
+import no.nav.data.catalog.backend.app.github.GithubConsumer;
+import no.nav.data.catalog.backend.app.github.GithubProperties;
+import no.nav.data.catalog.backend.app.github.GithubWebhooksController;
 import no.nav.data.catalog.backend.app.github.domain.RepoModification;
 import no.nav.data.catalog.backend.app.informationtype.InformationType;
 import no.nav.data.catalog.backend.app.informationtype.InformationTypeRepository;
