--- data-catalog-backend-test/data-catalog-backend-test-component/src/test/java/no/nav/data/catalog/backend/test/component/PostgresTestContainer.java
@@ -1,44 +1,0 @@
-package no.nav.data.catalog.backend.test.component;
-
-import org.springframework.boot.test.util.TestPropertyValues;
-import org.springframework.context.ApplicationContextInitializer;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.testcontainers.containers.PostgreSQLContainer;
-
-public class PostgresTestContainer extends PostgreSQLContainer<PostgresTestContainer> {
-
-    private static final String IMAGE_VERSION = "postgres:10.4";
-    private static PostgresTestContainer container;
-
-    private PostgresTestContainer() {
-        super(IMAGE_VERSION);
-    }
-
-    public static PostgresTestContainer getInstance() {
-        if (container == null) {
-            container = new PostgresTestContainer();
-        }
-        return container;
-    }
-
-    @Override
-    public void start() {
-        super.start();
-    }
-
-    @Override
-    public void stop() {
-        //do nothing, JVM handles shut down
-    }
-
-    public static class Initializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
-
-        public void initialize(ConfigurableApplicationContext configurableApplicationContext) {
-            TestPropertyValues.of(
-                    "spring.datasource.url=" + getInstance().getJdbcUrl(),
-                    "spring.datasource.username=" + getInstance().getUsername(),
-                    "spring.datasource.password=" + getInstance().getPassword()
-            ).applyTo(configurableApplicationContext.getEnvironment());
-        }
-    }
-}
