+++ data-catalog-backend-test/data-catalog-backend-test-integration/src/test/java/no/nav/data/catalog/backend/app/IntegrationTestConfig.java
@@ -1,0 +1,7 @@
+package no.nav.data.catalog.backend.app;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackageClasses = {IntegrationTestConfig.class})
+public class IntegrationTestConfig {
+}
