--- data-catalog-backend-test/data-catalog-backend-test-integration/src/test/java/no/nav/data/catalog/backend/test/integration/IntegrationTestConfig.java
@@ -1,7 +1,0 @@
-package no.nav.data.catalog.backend.test.integration;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication(scanBasePackageClasses = {IntegrationTestConfig.class})
-public class IntegrationTestConfig {
-}
